Adds a sixth JSON data structure.
Adds documentation to each method that generates a JSON data structure, so
that the script need not be run six times to determine which the desired
JSON data structure.

The structures have fetched a rename as well (method names, prefix 'xmlTo'):

* "All the Fields" -> JsonObject
* "Unnamed Array/Verbose" -> VerboseJsonArray
* "Named Array/Verbose" -> NamedVerboseJsonArray
* "Unnamed Array/Compressed" -> CompressedJsonArray
* "Named Array/Compressed" -> NamedCompressedJsonArray

Note that the "Compressed" arrays do not "minimalize" the file, in that
individual elements take up a minimum of 3 lines:

{
	"P1000": "Example code"
}

Minimalization can be affected by removing any leading or trailing whitespace
characters per line (as all text descriptions of a code are single-line), but
this functionality is not included at this time.